+++ src/main/java/com/jc/tm/Main.java
@@ -1,14 +1,10 @@
 package com.jc.tm;
 
-import com.sun.org.slf4j.internal.LoggerFactory;
-import lombok.extern.slf4j.Slf4j;
-
 /**
  * this main class for launch a whole application
  */
-@Slf4j
 public class Main {
     public static void main(String[] args) {
-        System.out.println("Main");
+        System.out.println("Hello world");
     }
 }
