+++ src/main/java/com/jc/tm/database/Status.java
@@ -1,0 +1,10 @@
+package com.jc.tm.database;
+
+/**
+ * status of for task
+ */
+public enum Status {
+    TODO,
+    IN_PROGRESS,
+    COMPLETE
+}
