--- src/main/java/database/EnumStatus.java
@@ -1,9 +1,0 @@
-package database;
-
-public class EnumStatus {
-    private enum Status{
-        ToDo,
-        InProgress,
-        Complate
-    }
-}
